+++ DirectX11_YooJaeJun/DirectX11_YooJaeJun/ObLine.cpp
@@ -17,8 +17,8 @@ void ObLine::Render()
         TransfVtx[i] = Vector2::Transform(vertex[i], W);
     }
 
-    MoveToEx(g_MemDC, TransfVtx[0].x, TransfVtx[0].y, NULL);
-    LineTo(g_MemDC, TransfVtx[1].x, TransfVtx[1].y);
+    MoveToEx(g_MemDC, (int)TransfVtx[0].x, (int)TransfVtx[0].y, NULL);
+    LineTo(g_MemDC, (int)TransfVtx[1].x, (int)TransfVtx[1].y);
 }
 
 /*
